refactor: extract test-kit from tests
